Bringing Master Fork up to date

+++ app/bundles/CoreBundle/IpLookup/AbstractLocalDataLookup.php
@@ -1,9 +1,11 @@
 <?php
-/**
- * @package     Mautic
- * @copyright   2015 Mautic Contributors. All rights reserved.
+
+/*
+ * @copyright   2015 Mautic Contributors. All rights reserved
  * @author      Mautic
+ *
  * @link        http://mautic.org
+ *
  * @license     GNU/GPLv3 http://www.gnu.org/licenses/gpl-3.0.html
  */
 
@@ -12,19 +14,19 @@ namespace Mautic\CoreBundle\IpLookup;
 use Joomla\Http\HttpFactory;
 
 /**
- * Class AbstractLocalDataLookup
+ * Class AbstractLocalDataLookup.
  */
 abstract class AbstractLocalDataLookup extends AbstractLookup implements IpLookupFormInterface
 {
     /**
-     * Path to the local data store
+     * Path to the local data store.
      *
      * @return string
      */
     abstract public function getLocalDataStoreFilepath();
 
     /**
-     * Return the URL to manually download
+     * Return the URL to manually download.
      *
      * @return string
      */
@@ -45,11 +47,11 @@ abstract class AbstractLocalDataLookup extends AbstractLookup implements IpLooku
      */
     public function getConfigFormThemes()
     {
-        return array();
+        return [];
     }
 
     /**
-     * Download remote data store
+     * Download remote data store.
      *
      * Used by the mautic:iplookup:update_data command and form fetch button (if applicable) to update local IP data stores
      *
@@ -75,13 +77,17 @@ abstract class AbstractLocalDataLookup extends AbstractLookup implements IpLooku
             $success = false;
 
             switch (true) {
-                case ($localTargetExt === $tempExt):
+                case $localTargetExt === $tempExt:
                     $success = (bool) file_put_contents($localTarget, $data->body);
 
                     break;
 
-                case ('gz' == $tempExt):
-                    if (function_exists('gzdecode')) {
+                case 'gz' == $tempExt:
+                    $memLimit = $this->sizeInByte(ini_get('memory_limit'));
+                    $freeMem  = $memLimit - memory_get_peak_usage();
+                    //check whether there is enough memory to handle large iplookp DB
+                    // or will throw iplookup exception
+                    if (function_exists('gzdecode') && strlen($data->body) < ($freeMem / 3)) {
                         $success = (bool) file_put_contents($localTarget, gzdecode($data->body));
                     } elseif (function_exists('gzopen')) {
                         if (file_put_contents($tempTarget, $data->body)) {
@@ -100,7 +106,7 @@ abstract class AbstractLocalDataLookup extends AbstractLookup implements IpLooku
 
                     break;
 
-                case ('zip' == $tempExt):
+                case 'zip' == $tempExt:
                     file_put_contents($tempTarget, $data->body);
 
                     $zipper = new \ZipArchive();
@@ -121,7 +127,7 @@ abstract class AbstractLocalDataLookup extends AbstractLookup implements IpLooku
     }
 
     /**
-     * Get the common directory for data
+     * Get the common directory for data.
      *
      * @return null|string
      */
@@ -137,5 +143,18 @@ abstract class AbstractLocalDataLookup extends AbstractLookup implements IpLooku
         }
 
         return null;
+    }
+
+    protected function sizeInByte($size)
+    {
+        $data = (int) substr($size, 0, -1);
+        switch (strtoupper(substr($size, -1))) {
+            case 'K':
+                return $data * 1024;
+            case 'M':
+                return $data * 1024 * 1024;
+            case 'G':
+                return $data * 1024 * 1024 * 1024;
+        }
     }
 }
